lock database when performing write operations

+++ server/routes/admin/upload.php
@@ -105,11 +105,19 @@
       }
     }
   }
+
+  if (isDatabaseLocked()) {
+    http_response_code(503);
+    echo json_encode(array("error" => "Database is locked by another process. Please try again later. Note if you think the database is incorrectly locked then it will automatically unlock in 20 minutes."));
+    return;
+  }
+  lockDatabase();
   
   $uploadStartTime = time();
   ingestSpreadsheet($newFilePath, basename($newFilePath), 0);
   deleteOldCrknData('filename', $uploadStartTime, basename($newFilePath));
 
+  unlockDatabase();
   $serverFiles = getXLSXFiles('../../PAR-files/');
   echo json_encode(array("files" => $serverFiles));
 ?>
